+++ src/main/java/app/bean/CourseBean.java
@@ -40,7 +40,7 @@ public class CourseBean {
         return course;
     }
 
-    public boolean delete(Integer id){
+    public boolean delete(Long id){
         if (id > 0 ) {
             auditTrailEvent.fire(new AuditTrail("Course Deleted, ID: " + id));
             courseDao.delete(id);
